backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Vector Database
pinecone-client==3.0.2

# ML and Embeddings
transformers==4.36.2
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
sentence-transformers==2.3.1

# LLM
groq==0.4.2

# HTTP Client 
httpx==0.27.0
httpcore==1.0.5

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.3
numpy==1.24.3
requests==2.31.0

